+++ examples/buggy/elixir/broken.ex
@@ -1,0 +1,5 @@
+defmodule Broken do
+  def run(name) do
+    IO.puts("hello #{name}"
+  end
+end