+++ examples/buggy/r/broken.R
@@ -1,0 +1,7 @@
+broken <- function(x) {
+  if (x > 0) {
+    print("positive")
+  else {
+    print("negative")
+  }
+}